added ablation study data generator

+++ opts.py
@@ -7,7 +7,7 @@ def fill_config(args):
     args.dec_ninp = args.nhid * 3 if args.title else args.nhid * 2
     args.fnames = [args.train_file, args.valid_file, args.test_file]
 
-    if not os.path.isfile(args.train_file): download()
+    download()
     return args
 
 def vocab_config(args, ent_vocab, rel_vocab, text_vocab, ent_text_vocab, title_vocab):
@@ -41,11 +41,11 @@ def get_args():
     args.add_argument('--drop', default=0.1, type=float)
     args.add_argument('--lp', default=1.1, type=float)
     args.add_argument('--graph_enc', default='gtrans', type=str)
-    args.add_argument('--train_file', default='data/agenda/unprocessed.train.json',
+    args.add_argument('--train_file', default='data/agenda/train.json',
                       type=str)
-    args.add_argument('--valid_file', default='data/agenda/unprocessed.valid.json',
+    args.add_argument('--valid_file', default='data/agenda/valid.json',
                       type=str)
-    args.add_argument('--test_file', default='data/agenda/unprocessed.test.json',
+    args.add_argument('--test_file', default='data/agenda/test.json',
                       type=str)
     args.add_argument('--save_dataset', default='data.pickle', type=str)
     args.add_argument('--save_model', default='saved_model.pt', type=str)
@@ -57,22 +57,29 @@ def get_args():
 
 def download():
 
-    print('[Info] Downloading AGENDA data...')
-    cmd = \
-        'curl -O https://raw.githubusercontent.com/rikdz/GraphWriter/master/data/unprocessed.tar.gz; ' \
-        'mkdir -p data/agenda; ' \
-        'rm data/agenda/* 2>/dev/null; ' \
-        'tar -C data/agenda -xvf unprocessed.tar.gz; ' \
-        'mv data/agenda/unprocessed.val.json data/agenda/unprocessed.valid.json; ' \
-        'rm unprocessed.tar.gz'
-    os.system(cmd)
+    if not os.path.isfile('data/agenda/train.json'):
+        print('[Info] Downloading AGENDA data...')
+        cmd = \
+            'curl -O https://raw.githubusercontent.com/rikdz/GraphWriter/master/data/unprocessed.tar.gz; ' \
+            'mkdir -p data/agenda/raw; ' \
+            'rm data/agenda/raw/* 2>/dev/null; ' \
+            'tar -C data/agenda/raw -xvf unprocessed.tar.gz; ' \
+            'mv data/agenda/raw/unprocessed.train.json data/agenda/train.json; ' \
+            'mv data/agenda/raw/unprocessed.test.json data/agenda/test.json; ' \
+            'mv data/agenda/raw/unprocessed.val.json data/agenda/valid.json; ' \
+            'rm unprocessed.tar.gz; ' \
+            'rm -rf data/agenda/raw/; ' \
+            'python data/ablation/full_agenda.py '
+        os.system(cmd)
+
 
     files = [('eval/detokenizer.perl',
               'https://raw.githubusercontent.com/moses-smt/mosesdecoder/master/scripts/tokenizer/detokenizer.perl'),
              ('eval/multi-bleu-detok.perl',
               'https://raw.githubusercontent.com/moses-smt/mosesdecoder/master/scripts/generic/multi-bleu-detok.perl')]
-    os.system('mkdir eval')
-    for fname, url in files:
-        if not os.path.isfile(fname):
-            os.system('curl -o {fname} {url}'.format(fname=fname, url=url))
+    if not os.path.isfile(files[0][0]):
+        os.system('mkdir eval')
+        for fname, url in files:
+            if not os.path.isfile(fname):
+                os.system('curl -o {fname} {url}'.format(fname=fname, url=url))
